deploy webinar app
