+++ src/apps/webinar/.env.development
@@ -1,1 +1,3 @@
 NEXT_PUBLIC_AUTH_URL="https://auth.dev.kloudlite.io"
+NEXT_PUBLIC_Callback_URL="https://events.dev.kloudlite.io"
+NEXT_PUBLIC_REDIRECT_URL="https://auth.dev.kloudlite.io/login"
