+++ src/apps/webinar/src/app/components/join-webinar.tsx
@@ -1,9 +1,7 @@
 'use client';
 import { ArrowRightLg } from '@jengaicons/react';
-import axios from 'axios';
 //@ts-ignore
 import { Button } from 'kl-design-system/atoms/button';
-import { useEffect, useState } from 'react';
 
 type UesrData = {
     id: string,
@@ -13,56 +11,22 @@ type UesrData = {
     approved: boolean,
 }
 
-export const JoinWebinar = ({ userData }: { userData: UesrData }) => {
-
-    const [meetingStatus, setMeetingStatus] = useState<boolean>(false);
-
-    const getMeetingDetails = async () => {
-        const token = btoa(`${process.env.NEXT_PUBLIC_DYTE_ORG_ID}:${process.env.NEXT_PUBLIC_DYTE_API_KEY}`);
-        try {
-            const { data: { success, data } } = await axios.get(
-                `https://api.dyte.io/v2/meetings/${process.env.NEXT_PUBLIC_DYTE_MEETING_ID}`,
-                {
-                    headers: {
-                        Authorization: `Basic ${token}`,
-                    }
-                }
-            )
-            if (!success) {
-                throw new Error('Failed to get meeting details');
-            }
-            console.log("++++++>>>>>>>>>>>>>>", data);
-
-            setMeetingStatus(data.status === 'ACTIVE');
-        } catch (error) {
-            console.error(error);
-        }
-    }
-
-    useEffect(() => {
-        (async () => {
-            await getMeetingDetails();
-        })();
-    }, []);
-
+export const JoinWebinar = ({ userData, meetingStatus }: { userData: UesrData, meetingStatus: string }) => {
 
     return (
         <div
             className='flex flex-col items-stretch gap-3xl'>
-            {
-                meetingStatus && (
-                    <Button
-                        size="lg"
-                        variant="primary"
-                        content={<span className="bodyLg-medium">Join</span>}
-                        suffix={<ArrowRightLg />}
-                        block
-                        onClick={() => {
-                            window.location.href = `/pages/meeting?email=${userData.email}&name=${userData.name}&meetingId=${process.env.NEXT_PUBLIC_DYTE_MEETING_ID}`
-                        }}
-                    />
-                )
-            }
+            <Button
+                size="lg"
+                variant="primary"
+                content={<span className="bodyLg-medium">Join</span>}
+                suffix={<ArrowRightLg />}
+                disabled={meetingStatus !== 'ACTIVE'}
+                block
+                onClick={() => {
+                    window.location.href = `/pages/meeting?email=${userData.email}&name=${userData.name}&meetingId=${process.env.NEXT_PUBLIC_DYTE_MEETING_ID}`
+                }}
+            />
         </div>
     )
 }
