+++ src/design-system/build.js
@@ -1,6 +1,7 @@
 import { build } from 'esbuild';
 import { sassPlugin } from 'esbuild-sass-plugin';
-import p from './package.json' with { type: "json" };
+import p from './package.json' assert { type: 'json' };
+// import p from './folder/file.json' assert { type: 'json' };
 
 const { externalDependencies } = p;
